Add REUSE.toml files and missing licenses
REUSE.toml files are read by reuse to complement or override the
copyright and licensing information found in file.

The use of REUSE.toml files was introduced in REUSE version 3.1.0a1.
This reuse version is compatible with reuse specification
version 3.2 [1].
With this commit's files,
* The SPDX document generated by reuse spdx conforms to SPDX 2.3,
* The reuse lint command reports that the Qt project is reuse compliant.

In order to be reuse compliant all the licenses referenced in file
or within a REUSE.toml file must be present in the LICENSES
directory at the base of the module.
The missing licenses are added.

[1]: https://reuse.software/spec-3.2/

Task-number: QTBUG-124453
Task-number: QTBUG-125211
Change-Id: If07497860c819b00169a349edde641a29a5ef259
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit e776c032ed158862772c6995b113501d8b426e2c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>