UpdateWorldTransformJob: fix compilation with C++23 GCC 14 unity-builds
GCC doesn't like the use of a struct in the unnamed namespace in a
struct that's outside said namespace:

  updateworldtransformjob.cpp:60:39: error: ‘Qt3DRender::Render::UpdateWorldTransformJobPrivate’ has a field ‘QList<Qt3DRender::Render::{anonymous}::TransformUpdate> Qt3DRender::Render::UpdateWorldTransformJobPrivate::m_updatedTransforms’ whose type has internal linkage [-Werror=subobject-linkage]
   60 | class Q_3DRENDERSHARED_PRIVATE_EXPORT UpdateWorldTransformJobPrivate : public Qt3DCore::QAspectJobPrivate
      |                                       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

I don't know whether it's GCC 14, C++23, or unity-builds that cause
this warning, and I don't have the resources to do a differential
analysis on it, sorry, so picking to all relevant branches.

Fix by moving the struct out of the unnamed namespace. There is no
other use of the name anywhere in qt3d, so the unnamed namespace
wasn't necessary to begin with.

Amends 48ff8579fe5643e2b194669bff2c12e69e00053e.

Pick-to: 6.9 6.8 6.5 5.15
Change-Id: I448da205e27f1fe3acc26db515d00ba2cddc53df
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>